Allow application to specify the metadata retransmitted with the stream
Currently, the draft only allows the data in the beginning of the stream
to be made reliable.  This PR instead proposes to let application
specify what is being reset.  This has the drawback of limiting the
amount of reset-attached metadata we can transmit to ~1200 bytes, but it
also allows us to address issue
ietf-wg-webtrans/draft-ietf-webtrans-http3#90, and since the only things
we currently need to be reliable are the session ID and the error code,
this should not be the problem.