Move and rephrase the clause about zero Reliable Size
So part of my issue as a reader is the second paragraph is focused on detailing one
specific edge case of RESET_STREAM_AT. It's good to cover the edge case but
I find it distracting to the primary use case. Therefore, propose moving it to the end
of the section. At the same time, reword things slightly so it is clear this is
just a special clause due to zero causing an effect like REST_STREAM.

Attempts to fix #53